refactor: introduce subnamespaces in OpenAPI project
